WIP: Introduce TypeScript and type Calypso Block Editor iFrame bridge
See prior work in #13117, #15390, and #15397

This PR introduces TypeScript into our Babel presets so that we can write in TypeScript and take advantage of the safety and development experience improvements that it offers.

I have updated the _iframe bridge_ for the block editor as a first transformation to not only demonstrate what an update looks like but also because the integrity of that module is relatively important and the introduction of types brings some immediate tangible benefits.